[#10] refactor: Message 위치 변경 및 BlogRepository 패키지 위치 변경
