+++ schedule/src/main/java/itcast/ai/dto/request/GPTSummaryRequest.java
@@ -1,7 +1,5 @@
 package itcast.ai.dto.request;
 
-import itcast.ai.Message;
-
 public record GPTSummaryRequest(
         String model,
         Message message,
