+++ schedule/src/main/java/itcast/ai/dto/request/Message.java
@@ -1,4 +1,4 @@
-package itcast.ai;
+package itcast.ai.dto.request;
 
 import lombok.AccessLevel;
 import lombok.Getter;
